+++ docs/releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## 2.0.0
+
+- Added `LoginJWTApiView` allowing the JWT Token being protected via `httpOnly=true` cookie
+  and refreshing the token via `middleware`. [Docs here](./views/auth.md)
+
 ## 1.0.3
 
 - Fixed typo in `PrefetchRelatedMixin`.